+++ react-native/ios/RealmReact/RealmReact.mm
@@ -79,8 +79,6 @@ extern "C" JSGlobalContextRef RealmReactGetJSGlobalContextForExecutor(id executo
 
 @implementation RealmReact {
     NSMutableDictionary *_eventHandlers;
-    __weak NSThread *_currentJSThread;
-    __weak NSRunLoop *_currentJSRunLoop;
 
 #if DEBUG
     GCDWebServer *_webServer;
@@ -213,23 +211,24 @@ RCT_REMAP_METHOD(emit, emitEvent:(NSString *)eventName withObject:(id)object) {
     [_webServer addDefaultHandlerForMethod:@"POST"
                               requestClass:[GCDWebServerDataRequest class]
                               processBlock:^GCDWebServerResponse *(GCDWebServerRequest* request) {
+        __typeof__(self) self = weakSelf;
+        RPCServer *rpcServer = self ? self->_rpcServer.get() : nullptr;
         GCDWebServerResponse *response;
+
         try {
-            // perform all realm ops on the main thread
-            __block NSData *responseData;
-            dispatch_sync(dispatch_get_main_queue(), ^{
-                RealmReact *self = weakSelf;
-                if (self) {
-                    if (_rpcServer) {
-                        json args = json::parse([[(GCDWebServerDataRequest *)request text] UTF8String]);
-                        std::string responseText = _rpcServer->perform_request(request.path.UTF8String, args).dump();
-                        responseData = [NSData dataWithBytes:responseText.c_str() length:responseText.length()];
-                        return;
-                    }
-                }
+            NSData *responseData;
+
+            if (rpcServer) {
+                json args = json::parse([[(GCDWebServerDataRequest *)request text] UTF8String]);
+                std::string responseText = rpcServer->perform_request(request.path.UTF8String, args).dump();
+
+                responseData = [NSData dataWithBytes:responseText.c_str() length:responseText.length()];
+            }
+            else {
                 // we have been deallocated
                 responseData = [NSData data];
-            });
+            }
+
             response = [[GCDWebServerDataResponse alloc] initWithData:responseData contentType:@"application/json"];
         }
         catch(std::exception &ex) {
@@ -260,17 +259,6 @@ RCT_REMAP_METHOD(emit, emitEvent:(NSString *)eventName withObject:(id)object) {
     // shutdown rpc if in chrome debug mode
     [self shutdownRPC];
 #endif
-
-    // block until JS thread exits
-    NSRunLoop *runLoop = _currentJSRunLoop;
-    if (runLoop) {
-        CFRunLoopStop([runLoop getCFRunLoop]);
-        while (_currentJSThread && !_currentJSThread.finished) {
-            [NSThread sleepForTimeInterval:0.01];
-        }
-    }
-
-    realm::_impl::RealmCoordinator::clear_all_caches();
 }
 
 - (void)dealloc {
@@ -295,15 +283,24 @@ RCT_REMAP_METHOD(emit, emitEvent:(NSString *)eventName withObject:(id)object) {
     }
     else {
         __weak __typeof__(self) weakSelf = self;
+        __weak __typeof__(executor) weakExecutor = executor;
 
         [executor executeBlockOnJavaScriptQueue:^{
             __typeof__(self) self = weakSelf;
-            if (!self) {
+            __typeof__(executor) executor = weakExecutor;
+            if (!self || !executor) {
                 return;
             }
 
-            self->_currentJSThread = [NSThread currentThread];
-            self->_currentJSRunLoop = [NSRunLoop currentRunLoop];
+            // Make sure the previous JS thread is completely finished before continuing.
+            static __weak NSThread *s_currentJSThread;
+            while (s_currentJSThread && !s_currentJSThread.finished) {
+                [NSThread sleepForTimeInterval:0.1];
+            }
+            s_currentJSThread = [NSThread currentThread];
+
+            // Close all cached Realms from the previous JS thread.
+            realm::_impl::RealmCoordinator::clear_all_caches();
 
             JSGlobalContextRef ctx = RealmReactGetJSGlobalContextForExecutor(executor, true);
             RJSInitializeInContext(ctx);